+++ app/Lib/Healthchecks.php
@@ -39,10 +39,30 @@ class Healthchecks {
     static public function environment() {
         $checks['environment']['phpVersion'] = (version_compare(PHP_VERSION, '5.2.8', '>='));
         $checks['environment']['pcre'] = (Validation::alphaNumeric('cakephp'));
-        $checks['environment']['tmpWritable'] = is_writable(TMP);
-        $checks['environment']['imgPublicWritable'] = is_writable(IMAGES . DS . 'public');
+		$checks['environment']['tmpWritable'] = self::_checkRecursiveDirectoryWritable(TMP);
+        $checks['environment']['imgPublicWritable'] = self::_checkRecursiveDirectoryWritable(IMAGES . 'public/');
         return $checks;
     }
+
+	/**
+	 * Check that a directory and its content are writable
+	 *
+	 * @param $path
+	 * @return boolean
+	 */
+	static private function _checkRecursiveDirectoryWritable($path) {
+		$iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator($path), RecursiveIteratorIterator::SELF_FIRST);
+		foreach ( $iterator as $name => $fileInfo ) {
+			if (in_array($fileInfo->getFilename(), ['.', '..', 'empty'])) {
+				continue;
+			}
+			if (!is_writable($name)) {
+				return false;
+			}
+		}
+
+		return true;
+	}
 
 /**
  * Return config file checks:
@@ -226,17 +246,24 @@ class Healthchecks {
          $checks['gpg']['gpgKeyNotDefault'] = (Configure::read('GPG.serverKey.fingerprint') != '2FC8945833C51946E937F9FED47B0811573EE67E');
 
          // Check keyring location is set and writable
-		 $checks['gpg']['gpgHome'] = (getenv('GNUPGHOME') !== false);
-		 $checks['gpg']['gpgHomeWritable'] = false;
-		 if($checks['gpg']['gpgHome']) {
-             $checks['gpg']['info']['gpgHome'] = getenv('GNUPGHOME');
-             $checks['gpg']['gpgHomeWritable'] = is_writable(getenv('GNUPGHOME'));
-         }
+		 // If the keyring location has been overridden.
+		 if (!empty(getenv('GNUPGHOME'))) {
+			 $checks['gpg']['info']['gpgHome'] = getenv('GNUPGHOME');
+		 }
+		 // Use the default user .gnupg location.
+		 else {
+			 $uid = posix_getuid();
+			 $user = posix_getpwuid($uid);
+			 $checks['gpg']['info']['gpgHome'] = $user['dir'] . '/.gnupg';
+		 }
+		 $checks['gpg']['gpgHome'] = file_exists($checks['gpg']['info']['gpgHome']);
+		 $checks['gpg']['gpgHomeWritable'] = is_writable($checks['gpg']['info']['gpgHome']);
 
          // Check key file exist and are readable
 		 $checks['gpg']['gpgKeyPublic'] = (Configure::read('GPG.serverKey.public') != null);
 		 $checks['gpg']['gpgKeyPublicReadable'] = is_readable(Configure::read('GPG.serverKey.public'));
 		 $checks['gpg']['gpgKeyPrivate'] = (Configure::read('GPG.serverKey.private') != null);
+		 $checks['gpg']['info']['gpgKeyPrivate'] = Configure::read('GPG.serverKey.private');
 		 $checks['gpg']['gpgKeyPrivateReadable'] = is_readable(Configure::read('GPG.serverKey.private'));
 
 		 // Check that the private key match the fingerprint
@@ -258,6 +285,17 @@ class Healthchecks {
 			 App::uses('Gpgkey', 'Model');
 			 $checks['gpg']['gpgKeyPublicEmail'] = Gpgkey::uidContainValidEmail($publicKeyInfo['uid']);
 		 }
+
+		 // Check that the private key is present in the keyring.
+		 $checks['gpg']['gpgKeyPrivateInKeyring'] = false;
+		 if ($checks['gpg']['gpgHome'] && Configure::read('GPG.serverKey.fingerprint')) {
+			 $gpg = new Passbolt\Gpg();
+			 $keyInfo = $gpg->getKeyInfoFromKeyring(Configure::read('GPG.serverKey.fingerprint'));
+			 if (!empty($keyInfo)) {
+				 $checks['gpg']['gpgKeyPrivateInKeyring'] = true;
+			 }
+		 }
+
 		 return $checks;
      }