+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - feature/PASSBOLT-2371: Force the charset of the cake_sessions table in utf8
+- feature/PASSBOLT-2325: As system administrator I shouldn't be able to execute passbolt CLI commands as root
+- feature/PASSBOLT-2397: As system administrator I should see in the healthcheck if app/tmp content and app/webroot/img/public content are writable
 
 ## [1.6.4] - 2017-08-31
 ### Fixed